refactor: extract reused code
