+++ lib/widgets/charts/dashboard_chart.dart
@@ -8,6 +8,7 @@ class DashboardChart extends StatelessWidget {
     this.overlay,
     this.topMargin = 0,
     this.transparent = false,
+    this.clipBehavior,
     super.key,
   });
 
@@ -17,6 +18,7 @@ class DashboardChart extends StatelessWidget {
   final double height;
   final double topMargin;
   final bool transparent;
+  final Clip? clipBehavior;
 
   @override
   Widget build(BuildContext context) {
@@ -34,6 +36,7 @@ class DashboardChart extends StatelessWidget {
               right: 10,
             ),
             child: Card(
+              clipBehavior: clipBehavior,
               shape: const RoundedRectangleBorder(
                 borderRadius: BorderRadius.all(Radius.circular(10)),
               ),
